Refactor:공간 댓글 관련 컴포넌트들 아토믹 디자인으로 정리
